Refactored for structure
