+++ lib/src/features/participants/participants_view.dart
@@ -2,14 +2,14 @@ import 'package:centaur_scores/src/app.dart';
 import 'package:centaur_scores/src/repository/repository.dart';
 import 'package:centaur_scores/src/mvvm/events/loading_event.dart';
 import 'package:centaur_scores/src/mvvm/observer.dart';
-import 'package:centaur_scores/src/participants/participants_viewmodel.dart';
-import 'package:centaur_scores/src/scores/scores_view.dart';
+import 'package:centaur_scores/src/features/participants/participants_viewmodel.dart';
+import 'package:centaur_scores/src/features/score_card/scores_view.dart';
 import 'package:centaur_scores/src/style/loading_screen.dart';
 import 'package:centaur_scores/src/style/style_helper.dart';
-import 'package:centaur_scores/src/syncwidget/score_sync_widget.dart';
+import 'package:centaur_scores/src/features/syncwidget/score_sync_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import '../model/match_model.dart';
+import '../../model/match_model.dart';
 import 'participant_editor.dart';
 
 class ParticipantsView extends StatelessWidget {
