+++ lib/src/features/score_card/scores_view.dart
@@ -1,7 +1,7 @@
 import 'package:centaur_scores/src/app.dart';
 import 'package:centaur_scores/src/repository/repository.dart';
-import 'package:centaur_scores/src/scores/score_entry_fullpage_widget.dart';
-import 'package:centaur_scores/src/syncwidget/score_sync_widget.dart';
+import 'package:centaur_scores/src/features/score_card/score_entry_fullpage_widget.dart';
+import 'package:centaur_scores/src/features/syncwidget/score_sync_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
