+++ lib/src/features/score_card/scores_viewmodel.dart
@@ -3,7 +3,7 @@ import 'package:centaur_scores/src/mvvm/events/loading_event.dart';
 import 'package:centaur_scores/src/mvvm/observer.dart';
 import 'package:centaur_scores/src/mvvm/viewmodel.dart';
 
-import '../model/match_model.dart';
+import '../../model/match_model.dart';
 
 class ScoresViewmodel extends EventViewModel {
   final MatchRepository _repository;
