+++ lib/src/features/score_card/single_participant_footer.dart
@@ -1,8 +1,8 @@
 import 'package:centaur_scores/src/style/style_helper.dart';
 import 'package:flutter/material.dart';
 
-import '../model/match_model.dart';
-import '../model/participant_model.dart';
+import '../../model/match_model.dart';
+import '../../model/participant_model.dart';
 
 class SingeParticipantFooter extends StatelessWidget {
   final MatchModel model;
