+++ lib/src/features/score_card/single_participant_header_line_two.dart
@@ -1,10 +1,10 @@
-import 'package:centaur_scores/src/participants/participants_view.dart';
+import 'package:centaur_scores/src/features/participants/participants_view.dart';
 import 'package:centaur_scores/src/style/style_helper.dart';
 import 'package:flutter/material.dart';
 
-import '../model/group_info.dart';
-import '../model/match_model.dart';
-import '../model/participant_model.dart';
+import '../../model/group_info.dart';
+import '../../model/match_model.dart';
+import '../../model/participant_model.dart';
 
 class SingleParticipantHeaderLineTwo extends StatelessWidget {
   final MatchModel model;
