+++ lib/src/features/score_card/single_participant_score_form.dart
@@ -1,10 +1,10 @@
-import 'package:centaur_scores/src/score_entry/score_entry_single_end.dart';
+import 'package:centaur_scores/src/features/score_entry/score_entry_single_end.dart';
 import 'package:centaur_scores/src/style/style_helper.dart';
-import 'package:centaur_scores/src/scores/scores_viewmodel.dart';
+import 'package:centaur_scores/src/features/score_card/scores_viewmodel.dart';
 import 'package:flutter/material.dart';
 
-import '../model/match_model.dart';
-import '../model/participant_model.dart';
+import '../../model/match_model.dart';
+import '../../model/participant_model.dart';
 
 class SingeParticipantScoreForm extends StatelessWidget {
   final MatchModel _model;
