+++ lib/src/features/score_entry/participant_navigation_box.dart
@@ -1,8 +1,8 @@
-import 'package:centaur_scores/src/score_entry/score_entry_single_end_viewmodel.dart';
+import 'package:centaur_scores/src/features/score_entry/score_entry_single_end_viewmodel.dart';
 import 'package:centaur_scores/src/style/style_helper.dart';
 import 'package:flutter/material.dart';
 
-import '../model/participant_model.dart';
+import '../../model/participant_model.dart';
 
 typedef GetParticipantFunction = ParticipantModel? Function();
 typedef GoToParticipantFunction = void Function();
