+++ lib/src/features/score_entry/score_entry_single_end.dart
@@ -2,18 +2,18 @@ import 'package:centaur_scores/src/app.dart';
 import 'package:centaur_scores/src/repository/repository.dart';
 import 'package:centaur_scores/src/mvvm/events/loading_event.dart';
 import 'package:centaur_scores/src/mvvm/observer.dart';
-import 'package:centaur_scores/src/score_entry/end_next.dart';
-import 'package:centaur_scores/src/score_entry/end_previous.dart';
-import 'package:centaur_scores/src/score_entry/score_entry_single_end_viewmodel.dart';
+import 'package:centaur_scores/src/features/score_entry/end_next.dart';
+import 'package:centaur_scores/src/features/score_entry/end_previous.dart';
+import 'package:centaur_scores/src/features/score_entry/score_entry_single_end_viewmodel.dart';
 import 'package:centaur_scores/src/style/loading_screen.dart';
 import 'package:centaur_scores/src/style/style_helper.dart';
-import 'package:centaur_scores/src/syncwidget/score_sync_widget.dart';
+import 'package:centaur_scores/src/features/syncwidget/score_sync_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
-import '../model/group_info.dart';
-import '../model/match_model.dart';
-import '../model/participant_model.dart';
+import '../../model/group_info.dart';
+import '../../model/match_model.dart';
+import '../../model/participant_model.dart';
 import 'score_keyboard.dart';
 import 'end_this.dart';
 import 'participant_navigation_box.dart';
