+++ lib/src/features/score_entry/score_keyboard.dart
@@ -1,10 +1,10 @@
-import 'package:centaur_scores/src/score_entry/score_entry_single_end_viewmodel.dart';
+import 'package:centaur_scores/src/features/score_entry/score_entry_single_end_viewmodel.dart';
 import 'package:centaur_scores/src/style/style_helper.dart';
 import 'package:flutter/material.dart';
 
-import '../model/match_model.dart';
-import '../model/participant_model.dart';
-import '../model/score_button_definition.dart';
+import '../../model/match_model.dart';
+import '../../model/participant_model.dart';
+import '../../model/score_button_definition.dart';
 
 class ScoreKeyboard extends StatelessWidget {
   final ScoresSingleEndViewmodel _viewModel;
